requirements.txt
mysql-connector-python==9.1.0
streamlit-option-menu==0.4.0
streamlit==1.41.1
python-dotenv==1.0.0